Remove the bundler version selector
I do like the idea of bundler being able to lock its own version, but
this implementation is not the right way to do it in my opinion.

I aim to reimplement this inside bundler in the near future, but it
needs to be done in a more user friendly way that properly informs the
user about what's going on, and never raises when not necessary.

But for now I think we should remove this code from rubygems because
this is causing more problems than benefits and it's being
counterproductive.